testsuite: Add tests for PR118149
A recent bugfix (eee2891312) for PR117830 also addressed PR118149.
This patch adds two test cases for PR118149.
These tests are different than other tests in that one of the
vec-perm selectors contains indices in descending order (1, 1, 0, 0),
which is the root cause for the ICE observed in PR118149.

	PR118149

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/pr118149-2.c: New test.
	* gcc.dg/tree-ssa/pr118149.c: New test.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>