avoid trying to set block in barriers [PR113506]
When we emit a sequence before a preexisting insn and naming a BB to
store in the insns, we will attempt to store the BB even in barriers
present in the sequence.

Barriers don't expect blocks, and rtl checking catches the problem.

When emitting after a preexisting insn, we skip the block setting in
barriers.  Change the before emitter to do so as well.


for  gcc/ChangeLog

	PR middle-end/113506
	* emit-rtl.cc (add_insn_before): Don't set the block of a
	barrier.

for  gcc/testsuite/ChangeLog

	PR middle-end/113506
	* gcc.target/riscv/pr113506.c: New.