c++: add ref checks in conversion code
While looking at another patch I noticed that on a few tests we were doing
nonsensical things like building a reference to a reference.  Make sure we
catch that sooner.  But let's be friendly in can_convert, since it doesn't
return a conversion that could be wrongly applied to a reference.

gcc/cp/ChangeLog:

	* call.cc (implicit_conversion): Check that FROM isn't a reference
	if we also got an EXPR argument.
	(convert_like_internal): Check that EXPR isn't a reference.
	(can_convert_arg): convert_from_reference if needed.