[RISC-V][PR target/115375] Fix expected dump output
Several months ago changes were made to the vectorizer which mucked up several
of the scan tests.  All but one of the cases in pr115375 have since been fixed.

The remaining failure seems to be primarily a debugging dump issue -- we're
still selecting the same lmul values.  This patch adjusts the dump scan
appropriately.

	PR target/115375

gcc/testsuite

	* gcc.dg/vect/costmodel/riscv/rvv/dynamic-lmul8-9.c: Adjust expected
	output.