bpf: install a wrapping stdint.h for bpf-none-unknown targets
This patch makes GCC to wrap the host's stdint.h header when operating
in the default hosted/non-freestanding mode.  It is thus up to the
user to provide a suitable standard header unless -ffreestanding gets
explicitly passed in the compilation line.

A few tests that use stdint.h are adapted to explicitly pass
-ffreestanding.

Tested in a x86_64-linux-gnu host with target bpf-unknown-none.

gcc/ChangeLog:

	* config.gcc: install a wrapping stdint.h in bpf targets.

gcc/testsuite/ChangeLog:

	* gcc.target/bpf/diag-funargs-2.c: Use -ffreestanding.
	* gcc.target/bpf/helper-skb-ancestor-cgroup-id.c: Likewise.
	* gcc.target/bpf/helper-bind.c: Likewise.
	* gcc.target/bpf/diag-funargs-3.c: Likewise.