testsuite: Replace MMIX-specific adjustments with TARGET_CALLEE_COPIES-adjustments
With the dump now emitting "privatized symbols" in the default
"%s.%lu" format also for MMIX, there's still a difference for MMIX.
This time it's because numbers have changed (copies introduced before
this point) because it has TARGET_CALLEE_COPIES yielding true.
Redundant copies may have been elided at this point, but the change
in name remains.

Since that's true for other targets too, an obvious change is to
generalize the tested patterns to include TARGET_CALLEE_COPIES-true
targets, as a brief inspection of the history of these tests shows
that the point of these tests lie not in whether copies have been done
but in the part of the pattern that match a constant.

Also fixed up a "." where there should have been a "\\.".

	* gcc.dg/tree-ssa/vector-4.c: Replace MMIX adjustments with
	TARGET_CALLEE_COPIES-agnostic adjustments.
	* gcc.dg/tree-ssa/forwprop-36.c: Ditto.  Correct pattern to match a
	literal ".".